requirements.txt
altgraph==0.17.4
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
gherkin-official==28.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pefile==2023.2.7
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pywin32-ctypes==0.2.2
setuptools==70.1.1
Werkzeug==3.0.3
